Refactor packaging names
