+++ java/src/code/desafios/model/CapitaoNaoInformadoException.java
@@ -1,4 +1,4 @@
-package code.desafios.exception;
+package code.desafios.model;
 
 public class CapitaoNaoInformadoException extends RuntimeException {
