+++ java/src/code/desafios/model/TimeNaoEncontradoException.java
@@ -1,4 +1,4 @@
-package code.desafios.exception;
+package code.desafios.model;
 
 public class TimeNaoEncontradoException extends RuntimeException {
